+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# Version 16.0.1
+
+## Bugfixes
+
+* Fixed dependency to techdivision/import-configuration to version 4.0.0
+
+## Features
+
+* None
+
 # Version 16.0.0
 
 ## Bugfixes